MODBULKOPS-4 Used pojo instead of entity for Job
