+++ src/main/java/org/folio/bulkops/client/DataExportClient.java
@@ -1,6 +1,6 @@
 package org.folio.bulkops.client;
 
-import org.folio.bulkops.domain.entity.Job;
+import org.folio.bulkops.domain.pojo.Job;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -9,8 +9,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.UUID;
-
-import static java.util.Objects.nonNull;
 
 @FeignClient
 public interface DataExportClient {
@@ -28,12 +26,6 @@ public interface DataExportClient {
   String startJob(@PathVariable UUID jobId);
 
   default String deleteFile(UUID jobId) {
-    var job = getJob(jobId);
-    if (nonNull(job)) {
-      var files = job.getFiles();
-      // TODO
-    }
-    // TODO
     return null;
   }
 }
